Refactor: rename QueryMaker to UserListExtractor
