+++ recommendator/src/main/java/com/lampirg/recommendator/anidb/mal/MalCommunicator.java
@@ -1,7 +1,7 @@
 package com.lampirg.recommendator.anidb.mal;
 
 import com.lampirg.recommendator.anidb.AnimeSiteCommunicator;
-import com.lampirg.recommendator.anidb.mal.querymaker.UserListExtractor;
+import com.lampirg.recommendator.anidb.mal.listextractor.UserListExtractor;
 import com.lampirg.recommendator.anidb.mal.titlemapper.TitleMapper;
 import com.lampirg.recommendator.anidb.model.AnimeRecommendation;
 import com.lampirg.recommendator.anidb.model.AnimeTitle;
